Moved RepositoryAdapter into standalone package.
